Add remote key tracker

+++ app/src/main/java/com/anilyilmaz/n11casestudy/core/database/entity/UserEntity.kt
@@ -5,7 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "users")
 data class UserEntity (
-    @PrimaryKey val id: Int,
+    @PrimaryKey
+    val id: Long,
     val login: String,
     val nodeId: String,
     val avatarUrl: String,
